refactor: dimmer 영역 클릭 시 꺼지는 로직 훅으로 분리
